requirements.txt
attrs==24.2.0
awsebcli==3.21.0
blessed==1.20.0
blinker==1.8.2
boto3==1.35.44
botocore==1.35.44
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-WTF==1.2.1
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathspec==0.10.1
pillow==11.0.0
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
semantic-version==2.10.0
sentence-transformers==3.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.45.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.2.0
XlsxWriter==3.2.0
